[tune/structure] Introduce execution package (#26015)
Execution-specific packages are moved to tune.execution.
