+++ python/ray/tune/__init__.py
@@ -38,7 +38,7 @@ from ray.tune.sample import (
 )
 from ray.tune.suggest import create_searcher
 from ray.tune.schedulers import create_scheduler
-from ray.tune.utils.placement_groups import PlacementGroupFactory
+from ray.tune.execution.placement_groups import PlacementGroupFactory
 from ray.tune.utils.trainable import with_parameters
 
 from ray._private.usage import usage_lib
