+++ python/ray/tune/analysis/experiment_analysis.py
@@ -30,7 +30,7 @@ from ray.tune.result import (
     TRAINING_ITERATION,
 )
 from ray.tune.trial import Trial
-from ray.tune.trial_runner import (
+from ray.tune.execution.trial_runner import (
     find_newest_experiment_checkpoint,
     load_trial_from_checkpoint,
 )
