+++ python/ray/tune/schedulers/async_hyperband.py
@@ -4,7 +4,7 @@ from typing import Dict, Optional, Union
 import numpy as np
 import pickle
 
-from ray.tune import trial_runner
+from ray.tune.execution import trial_runner
 from ray.tune.result import DEFAULT_METRIC
 from ray.tune.schedulers.trial_scheduler import FIFOScheduler, TrialScheduler
 from ray.tune.trial import Trial
