+++ python/ray/tune/schedulers/resource_changing_scheduler.py
@@ -7,11 +7,11 @@ import pickle
 import warnings
 
 from ray.util.annotations import PublicAPI
-from ray.tune import trial_runner
+from ray.tune.execution import trial_runner
 from ray.tune.resources import Resources
 from ray.tune.schedulers.trial_scheduler import FIFOScheduler, TrialScheduler
 from ray.tune.trial import Trial
-from ray.tune.utils.placement_groups import PlacementGroupFactory
+from ray.tune.execution.placement_groups import PlacementGroupFactory
 
 logger = logging.getLogger(__name__)
