+++ python/ray/tune/schedulers/trial_scheduler.py
@@ -1,6 +1,6 @@
 from typing import Dict, Optional
 
-from ray.tune import trial_runner
+from ray.tune.execution import trial_runner
 from ray.tune.result import DEFAULT_METRIC
 from ray.tune.trial import Trial
 from ray.util.annotations import DeveloperAPI, PublicAPI
