+++ python/ray/tune/tests/test_checkpoint_manager.py
@@ -7,7 +7,7 @@ import unittest
 from unittest.mock import patch
 
 from ray.tune.result import TRAINING_ITERATION
-from ray.tune.checkpoint_manager import _CheckpointManager
+from ray.tune.execution.checkpoint_manager import _CheckpointManager
 from ray.util.ml_utils.checkpoint_manager import (
     _TrackedCheckpoint,
     logger,
