+++ python/ray/tune/tests/test_trial_runner_callbacks.py
@@ -11,7 +11,7 @@ import ray
 from ray import tune
 from ray.rllib import _register_all
 from ray.tune.logger import DEFAULT_LOGGERS, LoggerCallback, LegacyLoggerCallback
-from ray.tune.ray_trial_executor import (
+from ray.tune.execution.ray_trial_executor import (
     _ExecutorEvent,
     _ExecutorEventType,
     RayTrialExecutor,
@@ -21,7 +21,7 @@ from ray.tune.syncer import SyncConfig, SyncerCallback
 
 from ray.tune.callback import warnings
 from ray.tune.trial import Trial
-from ray.tune.trial_runner import TrialRunner
+from ray.tune.execution.trial_runner import TrialRunner
 from ray.tune import Callback
 from ray.tune.utils.callback import create_default_callbacks
 from ray.tune.experiment import Experiment
