+++ python/ray/tune/tests/test_trial_runner_pg.py
@@ -6,11 +6,11 @@ import unittest
 
 import ray
 from ray import tune
-from ray.tune.ray_trial_executor import RayTrialExecutor
+from ray.tune.execution.ray_trial_executor import RayTrialExecutor
 from ray.tune.trial import Trial
 from ray.tune import Callback
-from ray.tune.trial_runner import TrialRunner
-from ray.tune.utils.placement_groups import PlacementGroupFactory
+from ray.tune.execution.trial_runner import TrialRunner
+from ray.tune.execution.placement_groups import PlacementGroupFactory
 from ray.util import placement_group_table
 from ray.cluster_utils import Cluster
 from ray.rllib import _register_all
