+++ python/ray/tune/tests/test_trial_scheduler_pbt.py
@@ -12,8 +12,8 @@ import ray
 from ray import tune
 from ray.tune import Trainable
 from ray.tune.trial import Trial
-from ray.tune.trial_runner import TrialRunner
-from ray.tune.ray_trial_executor import RayTrialExecutor
+from ray.tune.execution.trial_runner import TrialRunner
+from ray.tune.execution.ray_trial_executor import RayTrialExecutor
 from ray.tune.schedulers import PopulationBasedTraining
 from ray._private.test_utils import object_memory_usage
