+++ release/tune_tests/scalability_tests/workloads/test_result_throughput_cluster.py
@@ -16,7 +16,7 @@ import os
 
 import ray
 from ray import tune
-from ray.tune.cluster_info import is_ray_cluster
+from ray.tune.execution.cluster_info import is_ray_cluster
 
 from ray.tune.utils.release_test_util import timed_tune_run
