+++ rllib/algorithms/apex_dqn/apex_dqn.py
@@ -51,7 +51,7 @@ from ray.rllib.utils.typing import (
     ResultDict,
 )
 from ray.tune.trainable import Trainable
-from ray.tune.utils.placement_groups import PlacementGroupFactory
+from ray.tune.execution.placement_groups import PlacementGroupFactory
 from ray.util.ml_utils.dict import merge_dicts
